Changed operand type from Int to Double

+++ src/projects/calculator/Calculator.java
@@ -1,22 +1,22 @@
 package projects.calculator;
 
 public class Calculator {
-    private int num1;
-    private int num2;
+    private double num1;
+    private double num2;
 
-    public int getNum1(){
+    public double getNum1(){
         return num1;
     }
 
-    public int getNum2(){
+    public double getNum2(){
         return num2;
     }
 
-    public void setNum1(int num1){
+    public void setNum1(double num1){
         this.num1 = num1;
     }
 
-    public void setNum2(int num2){
+    public void setNum2(double num2){
         this.num2 = num2;
     }
 
@@ -29,19 +29,19 @@ public class Calculator {
         System.out.print("--> ");
     }
 
-    public int add(int num1, int num2){
+    public double add(double num1, double num2){
         return num1 + num2;
     }
 
-    public int subtract(int num1, int num2){
+    public double subtract(double num1, double num2){
         return num1 - num2;
     }
 
-    public int multiply(int num1, int num2){
+    public double multiply(double num1, double num2){
         return num1 * num2;
     }
 
-    public double div(int num1, int num2){
+    public double div(double num1, double num2){
         double result = (double) num1 / num2;
         return Math.round(result * 100) / 100.0;
     }
